+++ test/util.js
@@ -1,43 +1,48 @@
-const test = require('tape');
+const test = require('tape')
 
-const { dedent, flatten } = require('../util');
+const { dedent, flatten } = require('../util')
 
 test('util', (t) => {
-  t.plan(2);
-
   t.test('@dedent', (t) => {
     t.test('should work on simple es6 template string', (t) => {
-      let dedented = dedent(`
+      t.plan(1)
+
+      const dedented = dedent(`
         Usage
           $ foo <input>
 
         Commands:
           start, -s               Starts foo!
-      `);
-      t.equal(dedented, ' Usage\n   $ foo <input>\n\n Commands:\n   start, -s               Starts foo!\n      ');
-      t.end();
-    });
+      `)
+
+      t.equal(dedented, ' Usage\n   $ foo <input>\n\n Commands:\n   start, -s               Starts foo!\n      ')
+    })
 
     t.test('should work on mixed tabs and spaces', (t) => {
-      let dedented = dedent(`
+      t.plan(1)
+
+      const dedented = dedent(`
         \tUsage
         \t\t$ foo <input>
         \t
         \tCommands:
         \t\tstart, -s               Starts foo!
-      `);
-      t.equal(dedented, ' Usage\n     $ foo <input>\n \n Commands:\n     start, -s               Starts foo!\n      ');
-      t.end();
-    });
-  });
+      `)
+
+      t.equal(dedented, ' Usage\n     $ foo <input>\n \n Commands:\n     start, -s               Starts foo!\n      ')
+    })
+  })
 
   t.test('@flatten', (t) => {
     t.test('should only include alias and no shorthands', (t) => {
-      let map = flatten({
+      t.plan(1)
+
+      const map = flatten({
         start: {
           alias: ['s', 'vroom']
         }
-      });
+      })
+
       t.deepEqual(map, {
         start: {
           name: 'start'
@@ -48,17 +53,19 @@ test('util', (t) => {
         vroom: {
           name: 'start'
         }
-      });
-      t.end();
-    });
+      })
+    })
 
     t.test('should include alias and only default shorthand', (t) => {
-      let map = flatten({
+      t.plan(1)
+
+      const map = flatten({
         rainbow: {
           type: 'boolean',
           alias: 'r'
         }
-      }, true);
+      }, true)
+
       t.deepEqual(map, {
         rainbow: {
           name: 'rainbow',
@@ -76,17 +83,19 @@ test('util', (t) => {
           name: 'rainbow',
           type: 'boolean'
         }
-      });
-      t.end();
-    });
+      })
+    })
 
     t.test('should include alias and all shorthands', (t) => {
-      let map = flatten({
+      t.plan(1)
+
+      const map = flatten({
         rainbow: {
           type: 'boolean',
           alias: 'r'
         }
-      }, true, true);
+      }, true, true)
+
       t.deepEqual(map, {
         rainbow: {
           name: 'rainbow',
@@ -112,17 +121,19 @@ test('util', (t) => {
           name: 'rainbow',
           type: 'boolean'
         }
-      });
-      t.end();
-    });
+      })
+    })
 
     t.test('should include alias and all shorthands with nested alias', (t) => {
-      let map = flatten({
+      t.plan(1)
+
+      const map = flatten({
         rainbow: {
           type: 'boolean',
           alias: ['r', 'rain', 'bow']
         }
-      }, true, true);
+      }, true, true)
+
       t.deepEqual(map, {
         rainbow: {
           name: 'rainbow',
@@ -172,9 +183,7 @@ test('util', (t) => {
           name: 'rainbow',
           type: 'boolean'
         }
-      });
-      t.end();
-    });
-  });
-
-});
+      })
+    })
+  })
+})